chore: split runner file and organize
